requirements.txt
numpy
opencv-python
python-xlib
